AsyncAurora: code cleanup
