+++ etc/async-aurora/src/lib.rs
@@ -1,5 +1,7 @@
 mod engine_types;
 use crate::engine_types::*;
+mod async_promise;
+use crate::async_promise::*;
 use near_sdk::assert_self;
 use near_sdk::borsh::{self, BorshDeserialize, BorshSerialize};
 use near_sdk::{env, ext_contract, near_bindgen, AccountId, PanicOnDefault, Promise};
@@ -7,16 +9,6 @@ use near_sdk::{env, ext_contract, near_bindgen, AccountId, PanicOnDefault, Promi
 #[near_bindgen]
 #[derive(BorshDeserialize, BorshSerialize, PanicOnDefault)]
 pub struct AsyncAurora {
-}
-
-#[ext_contract(ext_bridge_token_factory)]
-pub trait ExtBridgeTokenFactory {
-    #[result_serializer(borsh)]
-    fn finish_withdraw(
-        &self,
-        #[serializer(borsh)] amount: Balance,
-        #[serializer(borsh)] recipient: AccountId,
-    ) -> Promise;
 }
 
 #[ext_contract(ext_aurora)]
@@ -36,62 +28,6 @@ pub trait ExtAsyncAurora {
         #[serializer(borsh)]
         result: SubmitResult,
     ) -> Promise;
-}
-
-const ERR_INVALID_PROMISE: &str = "Invalid promise format";
-
-enum CombinatorType {
-    And,
-    Then,
-}
-
-struct CombinatorDescription {
-    promise_index: u8,
-    combinator_type: CombinatorType,
-}
-
-struct PromiseDescription {
-    target: AccountId,
-    method_name: Vec<u8>,
-    arguments: Vec<u8>,
-    gas: u64,
-    combinator: Option<CombinatorDescription>,
-}
-
-fn parse_promises(input: String) -> Vec<PromiseDescription> {
-    let striped_input = input.strip_prefix("promises:").expect(ERR_INVALID_PROMISE);
-    let iter_list = striped_input.split("##");
-    let mut result = Vec::new();
-
-    for item in iter_list {
-        let mut iter = item.split('#');
-        let target = iter.next().expect(ERR_INVALID_PROMISE).into();
-        let method_name = iter.next().expect(ERR_INVALID_PROMISE).into();
-        let arguments = iter.next().expect(ERR_INVALID_PROMISE).into();
-        let gas = iter.next().expect(ERR_INVALID_PROMISE).parse::<u64>().expect(ERR_INVALID_PROMISE);
-        let mut combinator: Option<CombinatorDescription> = None;
-
-        if let Some(index_str) = iter.next() {
-            let promise_index = index_str.parse::<u8>().expect(ERR_INVALID_PROMISE);
-            let type_str = iter.next().expect(ERR_INVALID_PROMISE);
-            let combinator_type = if type_str == "&" {CombinatorType::And} else {CombinatorType::Then};
-            combinator = Some(CombinatorDescription {
-                promise_index,
-                combinator_type,
-            });
-        }
-
-        let promise_description = PromiseDescription {
-            target,
-            method_name,
-            arguments,
-            gas,
-            combinator: combinator,
-        };
-        result.push(promise_description);
-    }
-
-    result
 }
 
 #[near_bindgen]
